Remove deprecated ToolManager and migrate to BaseServer
This commit removes the ToolManager implementation and its associated tests, replacing it with functionality consolidated in BaseServer. Validation, tool management, and testing logic are now refactored to align with BaseServer's structure, improving code maintainability and reducing redundancy.
